Fix compilation

+++ runtime/src/main/cpp/Console.cpp
@@ -36,7 +36,7 @@ void Kotlin_io_Console_print(KString message) {
   utf8.reserve(message->count_);
   // Replace incorrect sequences with a default codepoint (see utf8::with_replacement::default_replacement)
   utf8::with_replacement::utf16to8(utf16, utf16 + message->count_, back_inserter(utf8));
-  konan::consoleWriteUtf8(utf8.c_str(), utf8.size());
+  konan::consoleWriteUtf8(utf8.c_str(), static_cast<uint32_t>(utf8.size()));
 }
 
 void Kotlin_io_Console_println(KString message) {
